VSCODE-NLP-297 Added help for special variables with dollar signs
