+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to the "nlp" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+### 1.53.0
+Added help for special variables with dollar signs
+
 ### 1.52.0
 Added analyzer processing queue for running multiple folders
